+++ src/imio/smartweb/core/browser/minisite/settings.py
@@ -3,18 +3,15 @@
 from Products.CMFPlone.interfaces.siteroot import IPloneSiteRoot
 from collective.instancebehavior import disable_behaviors
 from collective.instancebehavior import enable_behaviors
+from imio.smartweb.core.behaviors.minisite import IImioSmartwebMinisite
+from imio.smartweb.core.behaviors.minisite import IImioSmartwebMinisiteSettings
 from imio.smartweb.core.behaviors.subsite import IImioSmartwebSubsite
 from imio.smartweb.core.contents import IFolder
 from imio.smartweb.locales import SmartwebMessageFactory as _
 from plone import api
-from plone.app.layout.navigation.interfaces import INavigationRoot
 from Products.Five.browser import BrowserView
 from zope.interface import alsoProvides
 from zope.interface import noLongerProvides
-
-
-class IImioSmartwebMinisite(INavigationRoot):
-    """ """
 
 
 class MiniSiteSettings(BrowserView):
@@ -23,10 +20,12 @@ class MiniSiteSettings(BrowserView):
     def enable(self):
         enable_behaviors(
             self.context,
-            [IImioSmartwebSubsite.__identifier__],
-            [IImioSmartwebSubsite],
+            [IImioSmartwebMinisiteSettings.__identifier__],
+            [IImioSmartwebMinisiteSettings],
         )
         alsoProvides(self.context, IImioSmartwebMinisite)
+        self.context.exclude_from_nav = True
+        self.context.reindexObject(idxs=("object_provides", "exclude_from_nav"))
         api.portal.show_message(
             _(u"Minisite has been successfully activated"), self.request
         )
@@ -35,10 +34,12 @@ class MiniSiteSettings(BrowserView):
     def disable(self):
         disable_behaviors(
             self.context,
-            [IImioSmartwebSubsite.__identifier__],
-            [IImioSmartwebSubsite],
+            [IImioSmartwebMinisiteSettings.__identifier__],
+            [IImioSmartwebMinisiteSettings],
         )
         noLongerProvides(self.context, IImioSmartwebMinisite)
+        self.context.exclude_from_nav = False
+        self.context.reindexObject(idxs=("object_provides", "exclude_from_nav"))
         api.portal.show_message(_(u"Minisite has been disabled"), self.request)
         self.request.response.redirect(self.context.absolute_url())