+++ src/imio/smartweb/core/browser/subsite/settings.py
@@ -3,7 +3,7 @@
 from collective.instancebehavior import disable_behaviors
 from collective.instancebehavior import enable_behaviors
 from imio.smartweb.core.behaviors.subsite import IImioSmartwebSubsite
-from imio.smartweb.core.browser.minisite.settings import IImioSmartwebMinisite
+from imio.smartweb.core.behaviors.minisite import IImioSmartwebMinisite
 from imio.smartweb.core.contents import IFolder
 from imio.smartweb.locales import SmartwebMessageFactory as _
 from plone import api