+++ src/imio/smartweb/core/tests/test_minisite.py
@@ -1,20 +1,25 @@
 # -*- coding: utf-8 -*-
 
-from imio.smartweb.core.browser.minisite.settings import IImioSmartwebMinisite
-from imio.smartweb.core.behaviors.subsite import IImioSmartwebSubsite
-from imio.smartweb.core.testing import IMIO_SMARTWEB_CORE_INTEGRATION_TESTING
+from bs4 import BeautifulSoup
+from imio.smartweb.core.behaviors.minisite import IImioSmartwebMinisite
+from imio.smartweb.core.testing import IMIO_SMARTWEB_CORE_FUNCTIONAL_TESTING
 from imio.smartweb.core.testing import ImioSmartwebTestCase
 from plone import api
 from plone.app.layout.navigation.interfaces import INavigationRoot
 from plone.app.testing import TEST_USER_ID
+from plone.app.testing import TEST_USER_NAME
+from plone.app.testing import TEST_USER_PASSWORD
 from plone.app.testing import setRoles
 from plone.dexterity.content import ASSIGNABLE_CACHE_KEY
+from plone.testing.zope import Browser
 from zope.component import getMultiAdapter
 
+import transaction
 
-class MinisiteIntegrationTest(ImioSmartwebTestCase):
 
-    layer = IMIO_SMARTWEB_CORE_INTEGRATION_TESTING
+class MinisiteFunctionalTest(ImioSmartwebTestCase):
+
+    layer = IMIO_SMARTWEB_CORE_FUNCTIONAL_TESTING
 
     def setUp(self):
         """Custom shared utility setup for tests"""
@@ -39,7 +44,6 @@ class MinisiteIntegrationTest(ImioSmartwebTestCase):
         self.assertFalse(view.enabled)
 
         view.enable()
-        self.assertTrue(IImioSmartwebSubsite.providedBy(self.folder))
         self.assertTrue(IImioSmartwebMinisite.providedBy(self.folder))
         self.assertTrue(INavigationRoot.providedBy(self.folder))
 
@@ -47,7 +51,6 @@ class MinisiteIntegrationTest(ImioSmartwebTestCase):
         self.assertTrue(view.enabled)
 
         view.disable()
-        self.assertFalse(IImioSmartwebSubsite.providedBy(self.folder))
         self.assertFalse(IImioSmartwebMinisite.providedBy(self.folder))
         self.assertFalse(INavigationRoot.providedBy(self.folder))
         self.assertTrue(view.available)
@@ -58,6 +61,35 @@ class MinisiteIntegrationTest(ImioSmartwebTestCase):
         )
         subsite_view.enable()
         self.assertFalse(view.available)
+
+    def test_minisite_exclude_from_nav(self):
+        view = getMultiAdapter((self.folder, self.request), name="minisite_settings")
+        view.enable()
+        self.assertTrue(IImioSmartwebMinisite.providedBy(self.folder))
+        self.assertTrue(self.folder.exclude_from_nav)
+
+        transaction.commit()
+        browser = Browser(self.layer["app"])
+        browser.addHeader(
+            "Authorization",
+            "Basic %s:%s"
+            % (
+                TEST_USER_NAME,
+                TEST_USER_PASSWORD,
+            ),
+        )
+        browser.open("{}/edit".format(self.folder.absolute_url()))
+        content = browser.contents
+        soup = BeautifulSoup(content)
+        exclude_from_nav_widget = soup.find(
+            id="form-widgets-IExcludeFromNavigation-exclude_from_nav"
+        )
+        self.assertIsNone(exclude_from_nav_widget.find("input"))
+        children = [
+            c for c in exclude_from_nav_widget if not isinstance(c, str) or c.strip()
+        ]
+        self.assertEqual(len(children), 1)
+        self.assertEqual(children[0].text, "yes")
 
     def test_move_minisite_in_folder(self):
         view = getMultiAdapter((self.folder, self.request), name="minisite_settings")